Refactor - moved FeralDeceiverAbility to shared area
